move worker outside klee_web, add notes on the dependencies
