--- src/klee_web/worker/exceptions.py
@@ -1,2 +1,0 @@
-class KleeRunFailure(Exception):
-    pass
