--- src/klee_web/worker/tests/fixtures/simple/input.c
@@ -1,6 +1,0 @@
-#include <stdio.h>
-int main()
-{
-	printf("Hello world\n");
-	return 0;
-}
