--- src/klee_web/worker/tests/fixtures/symargs/expected.stdout
@@ -1,7 +1,0 @@
-(?:KLEE: .*?
-)+goodbye!
-hello!
-
-KLEE: done: total instructions = [0-9]+
-KLEE: done: completed paths = 2
-KLEE: done: generated tests = 2
