+++ src/worker/decorators.py
@@ -1,0 +1,9 @@
+def notify_on_entry(msg):
+    def decorator(f):
+        def wrapper(*args):
+            args[0].send_notification('notification', {'message': msg})
+            return f(*args)
+
+        return wrapper
+
+    return decorator
