+++ src/worker/mailer/dummy_mailer.py
@@ -1,0 +1,7 @@
+
+class DummyMailer():
+    def __init__(self, api_key=None):
+        pass
+
+    def send_mail(self, recipient, subject, email_body):
+        pass
