+++ src/worker/tests/fixtures/invalid_syntax/input.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+int main()
+{
+	printf("Hello world\n);
+	return 0;
+}
