+++ src/worker/tests/fixtures/simple/expected.stdout
@@ -1,0 +1,6 @@
+(?:KLEE: .*?
+)+Hello world
+
+KLEE: done: total instructions = [0-9]
+KLEE: done: completed paths = 1
+KLEE: done: generated tests = 1
