+++ src/worker/tests/fixtures/simple_unicode/expected.stdout
@@ -1,0 +1,6 @@
+(?:KLEE: .*?
+)+Zażółć gęślą jaźń
+
+KLEE: done: total instructions = [0-9]
+KLEE: done: completed paths = 1
+KLEE: done: generated tests = 1
