+++ src/worker/tests/fixtures/simple_unicode/input.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+int main(int argc, char** argv)
+{
+    printf("Zażółć gęślą jaźń\n");
+    return 0;
+}
