+++ src/worker/tests/fixtures/symargs/expected.stdout
@@ -1,0 +1,7 @@
+(?:KLEE: .*?
+)+goodbye!
+hello!
+
+KLEE: done: total instructions = [0-9]+
+KLEE: done: completed paths = 2
+KLEE: done: generated tests = 2
