+++ src/worker/tests/fixtures/symin/expected.stdout
@@ -1,0 +1,7 @@
+(?:KLEE: .*?
+)+(hello!
+|goodbye!
+){2}
+KLEE: done: total instructions = [0-9]+
+KLEE: done: completed paths = 2
+KLEE: done: generated tests = 2
